Initial move of ImageDisplay
