+++ rviz_common/include/rviz_common/image/ros_image_texture.hpp
