+++ rviz_common/src/rviz_common/image/ros_image_texture.cpp
